Functioning throttling and rate limit.

+++ db/migrate/20150620020316_create_games.rb
@@ -2,8 +2,8 @@ class CreateGames < ActiveRecord::Migration
   def change
     create_table :games do |t|
       t.text :gameData
-      t.integer :gameId, index: true
-      t.integer :summoner_id
+      t.bigint :gameId, index: true
+      t.bigint :summoner_id
       t.timestamps null: false
     end
     add_foreign_key :games, :summoners
